LPS-69043 Move command classes to their own package
