+++ modules/util/portal-tools-bundle-support/src/main/java/com/liferay/portal/tools/bundle/support/BundleSupport.java
@@ -16,6 +16,10 @@ package com.liferay.portal.tools.bundle.support;
 
 import com.beust.jcommander.JCommander;
 
+import com.liferay.portal.tools.bundle.support.commands.CommandClean;
+import com.liferay.portal.tools.bundle.support.commands.CommandDeploy;
+import com.liferay.portal.tools.bundle.support.commands.CommandDistBundle;
+import com.liferay.portal.tools.bundle.support.commands.CommandInitBundle;
 import com.liferay.portal.tools.bundle.support.internal.util.FileUtil;
 
 import java.io.File;
