+++ modules/util/portal-tools-bundle-support/src/main/java/com/liferay/portal/tools/bundle/support/commands/CommandDistBundle.java
@@ -12,12 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.tools.bundle.support;
+package com.liferay.portal.tools.bundle.support.commands;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 import com.beust.jcommander.converters.FileConverter;
 
+import com.liferay.portal.tools.bundle.support.BaseCommand;
 import com.liferay.portal.tools.bundle.support.internal.util.FileUtil;
 
 import java.io.File;
